product data class moved to vo package
