+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/dependencyinjection/FragmentsBuilderModule.kt
@@ -4,7 +4,7 @@ import com.cmpe352group4.buyo.ui.EmptyFragment
 import com.cmpe352group4.buyo.ui.ExampleFragment
 import com.cmpe352group4.buyo.ui.HomepageFragment
 import com.cmpe352group4.buyo.ui.navigationtabs.*
-import com.cmpe352group4.buyo.ui.productList.ProductList
+import com.cmpe352group4.buyo.ui.productList.ProductListFragment
 import com.cmpe352group4.buyo.ui.productDetail.ProductDetailCommentsFragment
 import com.cmpe352group4.buyo.ui.productDetail.ProductDetailContentFragment
 import dagger.Module
@@ -42,7 +42,7 @@ abstract class FragmentsBuilderModule {
     abstract fun contributeHomepageFragment(): HomepageFragment
 
     @ContributesAndroidInjector
-    abstract  fun contributeProductList(): ProductList
+    abstract  fun contributeProductListFragment(): ProductListFragment
 
     @ContributesAndroidInjector
     abstract  fun contributeProductDetailContentFragment(): ProductDetailContentFragment
