+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/ui/ExampleFragment.kt
@@ -7,8 +7,7 @@ import android.view.ViewGroup
 import com.cmpe352group4.buyo.R
 import com.cmpe352group4.buyo.base.BaseFragment
 import com.cmpe352group4.buyo.base.fragment_ops.TransactionType
-import com.cmpe352group4.buyo.ui.HomepageFragment
-import com.cmpe352group4.buyo.ui.productList.ProductList
+import com.cmpe352group4.buyo.ui.productList.ProductListFragment
 import com.cmpe352group4.buyo.ui.productDetail.ProductDetailContentFragment
 import kotlinx.android.synthetic.main.fragment_example.*
 
@@ -38,7 +37,7 @@ class ExampleFragment : BaseFragment() {
 
        btnProductList.setOnClickListener {
             navigationManager?.onReplace(
-                ProductList.newInstance(),
+                ProductListFragment.newInstance(),
                 TransactionType.Replace, true
             )
         }
