+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/ui/homepage/ProductAdapter.kt
@@ -7,6 +7,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.cmpe352group4.buyo.R
 import kotlinx.android.synthetic.main.square_product_item.view.*
+import com.cmpe352group4.buyo.vo.Product
 
 class ProductAdapter(
     var Products: MutableList<Product>,
