+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/ui/productList/ProductListFragment.kt
@@ -8,21 +8,18 @@ import android.widget.AdapterView
 import android.widget.SearchView
 import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.cmpe352group4.buyo.R
 import com.cmpe352group4.buyo.base.BaseFragment
 import com.cmpe352group4.buyo.base.fragment_ops.TransactionType
 import com.cmpe352group4.buyo.ui.EmptyFragment
 import kotlinx.android.synthetic.main.fragment_product_list.*
-import kotlinx.android.synthetic.main.item_product_list_recycler_view.*
-import java.time.*
-import java.util.*
+import com.cmpe352group4.buyo.vo.Product
 
 
-class ProductList : BaseFragment(){
+class ProductListFragment : BaseFragment(){
 
     companion object {
-        fun newInstance() = ProductList()
+        fun newInstance() = ProductListFragment()
     }
 
     override fun onCreateView(
@@ -56,12 +53,12 @@ class ProductList : BaseFragment(){
         var dummyComment = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus tempus sem tortor, id efficitur nisi feugiat eget. In ac odio sed nisl dapibus consequat. Praesent eu nulla at ipsum elementum varius et suscipit metus."
 
         var productsList = mutableListOf(
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName6", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020"),
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName5", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020" ),
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName4", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020" ),
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName3", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020" ),
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName2", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020" ),
-            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName1", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = 0.0, productReleaseDate = "01.01.2020")
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName6", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020"),
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName5", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020" ),
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName4", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020" ),
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName3", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020" ),
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName2", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020" ),
+            Product(productImage = "drawable/ic_launcher_background.xml", productInfo = dummyComment, productName = "MyItemName1", productID = 0, productNumComments = 0, productRate = 1.1, productPrice = "0.0", productReleaseDate = "01.01.2020")
         )
 
         val productListAdapter by lazy {
