+++ android/Buyo/app/src/main/res/layout/fragment_product_list.xml
@@ -8,7 +8,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.productList.ProductList">
+        tools:context=".ui.productList.ProductListFragment">
 
         <Button
             android:id="@+id/btnProductListBack"
